requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
pypdf
chromadb
faiss-cpu
groq
cassio
langchain-groq
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
